Implement rough board interface prototype Closes #3
- Auth implementation
