+++ apps/web/app/components/kiwinight-symbol.tsx
